Move back `networkIdSchema` to `zod.ts`
