+++ src/routes/token/ohlc/evm.ts
@@ -2,11 +2,10 @@ import { Hono } from 'hono';
 import { describeRoute } from 'hono-openapi';
 import { resolver, validator } from 'hono-openapi/zod';
 import { handleUsageQueryError, makeUsageQueryJson } from '../../../handleQuery.js';
-import { evmAddressSchema, statisticsSchema, paginationQuery, contractAddressSchema, intervalSchema, timestampSchema } from '../../../types/zod.js';
+import { evmAddressSchema, statisticsSchema, paginationQuery, contractAddressSchema, intervalSchema, timestampSchema, networkIdSchema } from '../../../types/zod.js';
 import { sqlQueries } from '../../../sql/index.js';
 import { z } from 'zod';
 import { DB_SUFFIX, DEFAULT_NETWORK_ID } from '../../../config.js';
-import { networkIdSchema } from '../../networks.js';
 import { stables } from '../../../inject/prices.tokens.js';
 
 const route = new Hono();
